# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudentStatus = typeof StudentStatus[keyof typeof StudentStatus];


export const StudentStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface Student {
  id: number;
  /**
     * @minLength 2
     * @maxLength 30
     */
  studentId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  lastName: string;
  /** @maxLength 255 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  department: string;
  /**
     * @minimum 1
     * @maximum 6
     */
  year: number;
  /**
     * @minLength 7
     * @maxLength 30
     */
  phone: string;
  enrollmentDate: string;
  status: StudentStatus;
  createdAt: string;
  updatedAt: string;
}

export interface StudentInput {
  /**
     * @minLength 2
     * @maxLength 30
     */
  studentId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  lastName: string;
  /** @maxLength 255 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  department: string;
  /**
     * @minimum 1
     * @maximum 6
     */
  year: number;
  /**
     * @minLength 7
     * @maxLength 30
     */
  phone: string;
  enrollmentDate: string;
  status?: StudentStatus;
}

export interface StudentUpdate {
  /**
     * @minLength 2
     * @maxLength 30
     */
  studentId?: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  firstName?: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  lastName?: string;
  /** @maxLength 255 */
  email?: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  department?: string;
  /**
     * @minimum 1
     * @maximum 6
     */
  year?: number;
  /**
     * @minLength 7
     * @maxLength 30
     */
  phone?: string;
  enrollmentDate?: string;
  status?: StudentStatus;
}

export interface StudentDepartmentCount {
  department: string;
  count: number;
}

export interface StudentsSummary {
  total: number;
  active: number;
  inactive: number;
  departments: StudentDepartmentCount[];
}

export interface Error {
  error: string;
}

export type ListStudentsParams = {
search?: string;
department?: string;
/**
 * @minimum 1
 * @maximum 6
 */
year?: number;
status?: StudentStatus;
};