# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentStatus } from './studentStatus';

export interface StudentInput {
  /**
     * @minLength 2
     * @maxLength 30
     */
  studentId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  lastName: string;
  /** @maxLength 255 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  department: string;
  /**
     * @minimum 1
     * @maximum 6
     */
  year: number;
  /**
     * @minLength 7
     * @maxLength 30
     */
  phone: string;
  enrollmentDate: Date;
  status?: StudentStatus;
}