# lib/api-zod/src/generated/types/studentStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type StudentStatus = typeof StudentStatus[keyof typeof StudentStatus];


export const StudentStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;