# lib/api-zod/src/generated/types/studentsSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentDepartmentCount } from './studentDepartmentCount';

export interface StudentsSummary {
  total: number;
  active: number;
  inactive: number;
  departments: StudentDepartmentCount[];
}